webapi/tct-video-html5-tests/video/video_mediaGroup_null.html

<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this lis
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work withou
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
 Xu,Liyu <liyux.xu@intel.com>

Revision history:
 Date               Author                                     Description
 04-28-2012         Xu,Liyu <liyux.xu@intel.com>
      case creation
-->

<!DOCTYPE html>
<html>
<head>
    <title>Video Test: video_mediagroup_null</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-html5-20121025/the-video-element.html#the-video-element" />
    <meta name="assert" content="Check if the value of attribute video.mediaGroup when set it null"/>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
</head>
<body>
    <video id="video1" controls="controls">
        Your browser does not support media elements.
    </video>
    <video id="video2" controls="controls">
        Your browser does not support HTML5 video.
    </video>
    <div id="log"></div>
    <script>
        test(function () {
            var video1 = document.getElementById('video1');
            var video2 = document.getElementById('video2');
            video1.mediaGroup = null;
            video2.mediaGroup = null;
            assert_equals(video1.mediaGroup,video2.mediaGroup,document.title);
        });
    </script>
  </body>
</html>
